requirements.txt
fastapi == 0.115.12
uvicorn == 0.29.0
python-multipart == 0.0.16
python-dotenv == 1.0.1